+++ package.json
@@ -1,7 +1,7 @@
 {
   "type": "module",
   "scripts": {
-    "build": "bash build.sh",
+    "build": "npx webpack",
     "lint": "npx eslint ./src --fix"
   },
   "devDependencies": {
@@ -9,6 +9,7 @@
     "@typescript-eslint/eslint-plugin": "^8.21.0",
     "@typescript-eslint/parser": "^8.21.0",
     "eslint": "^9.19.0",
+    "ts-loader": "^9.5.2",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.21.0",
     "uglify-js": "^3.19.3",
