+++ tsconfig.json
@@ -1,8 +1,12 @@
 {
   "compilerOptions": {
-    "target": "ES2017",
     "strict": true,
-    "outFile": "command-editor.min.js"
+    "outDir": "./dist/",
+    "noImplicitAny": true,
+    "module": "ES6",
+    "target": "ES2017",
+    "allowJs": true,
+    "moduleResolution": "node"
   },
   "include": ["src/**/*"],
 }
